+++ samples/DemoClient/App.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows;
+
+namespace DemoClient
+{
+    public partial class App : Application
+    {
+    }
+}